feat: valida se id existe no banco de dados

+++ src/routes.js
@@ -61,6 +61,13 @@ export const routes = [
       const { id } = req.params;
       const { title, description } = req.body;
 
+      const [task] = database.select("tasks", null, id);
+
+      if (!task)
+        return res
+          .writeHead(404)
+          .end(JSON.stringify({ message: "task not found!" }));
+
       if (!title)
         return res
           .writeHead(400)
@@ -71,13 +78,11 @@ export const routes = [
           .writeHead(400)
           .end(JSON.stringify({ message: "description is required!" }));
 
-      const task = {
+      database.update("tasks", id, {
         title,
         description,
         updated_at: new Date().toISOString(),
-      };
-
-      database.update("tasks", id, task);
+      });
 
       return res.writeHead(204).end();
     },
@@ -87,6 +92,13 @@ export const routes = [
     path: buildRoutePath("/tasks/:id"),
     handler: (req, res) => {
       const { id } = req.params;
+
+      const [task] = database.select("tasks", null, id);
+
+      if (!task)
+        return res
+          .writeHead(404)
+          .end(JSON.stringify({ message: "task not found!" }));
 
       database.delete("tasks", id);
